backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.76.0
pydantic==2.11.3